backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
feedparser>=6.0.11
pydantic>=2.6.0
python-dotenv>=1.0.1
anthropic>=0.25.0
networkx>=3.3
numpy>=1.26.0
jinja2>=3.1.3
markdown>=3.6
